Rename azure_monitor_log_profile(s) resource to azurerm_monitor_log_profile(s)
* Add deprecation warnings to azure_monitor_log_profile(s)
* Update docs to reference new resource names
* Update integration tests to use renamed resources
