+++ docs/resources/azurerm_monitor_log_profile.md.erb
@@ -1,20 +1,20 @@
 ---
-title: About the azure_monitor_log_profile Resource
+title: About the azurerm_monitor_log_profile Resource
 platform: azure
 ---
 
-# azure\_monitor\_log\_profile
+# azurerm\_monitor\_log\_profile
 
-Use the `azure_monitor_log_profile` InSpec audit resource to test properties
+Use the `azurerm_monitor_log_profile` InSpec audit resource to test properties
 of an Azure Monitor Log Profile.
 
 <br />
 
 ## Syntax
 
-An `azure_monitor_log_profile` resource block identifies a Log Profile by name.
+An `azurerm_monitor_log_profile` resource block identifies a Log Profile by name.
 
-    describe azure_monitor_log_profile(name: 'default') do
+    describe azurerm_monitor_log_profile(name: 'default') do
       ...
     end
 <br />
@@ -23,13 +23,13 @@ An `azure_monitor_log_profile` resource block identifies a Log Profile by name.
 
 ### Test that a Log Profile exists
 
-    describe azure_monitor_log_profile(name: 'default') do
+    describe azurerm_monitor_log_profile(name: 'default') do
       it { should exist }
     end
 
 ### Test that Log Profile retention is enabled
 
-    describe azure_monitor_log_profile(name: 'default') do
+    describe azurerm_monitor_log_profile(name: 'default') do
       its('retention_enabled') { should be true }
     end
 
@@ -43,7 +43,7 @@ An `azure_monitor_log_profile` resource block identifies a Log Profile by name.
 
 The name of the Log Profile.
 
-    describe azure_monitor_log_profile(name: 'default') do
+    describe azurerm_monitor_log_profile(name: 'default') do
       it { should exist }
     end
 
@@ -74,12 +74,12 @@ of available matchers, please visit our [Universal Matchers page](https://www.in
 The control will pass if the resource returns a result. Use `should_not` if you expect zero matches.
 
     #  If we expect 'default' to exist
-    describe azure_monitor_log_profile(name: 'default') do
+    describe azurerm_monitor_log_profile(name: 'default') do
       it { should exist }
     end
 
     # If we expect 'default' to not exist
-    describe azure_monitor_log_profile(name: 'default') do
+    describe azurerm_monitor_log_profile(name: 'default') do
       it { should_not exist }
     end
