+++ libraries/azure_monitor_log_profile.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require 'azurerm_resource'
+require 'azurerm_monitor_log_profile'
 
-class AzureMonitorLogProfile < AzurermResource
+class AzureMonitorLogProfile < AzurermMonitorLogProfile
   name 'azure_monitor_log_profile'
-  desc 'Verifies settings for a Azure Monitor Log Profile'
+  desc '[DEPRECATED] Please uze azurerm_monitor_log_profile'
   example <<-EXAMPLE
     describe azure_monitor_log_profile(name: 'default') do
       it { should exist }
@@ -13,40 +13,8 @@ class AzureMonitorLogProfile < AzurermResource
     end
   EXAMPLE
 
-  ATTRS = {
-    name:              'name',
-    id:                'id',
-    retention_policy:  'retentionPolicy',
-    retention_days:    'retentionPolicyDays',
-    retention_enabled: 'retentionPolicyEnabled',
-  }.freeze
-
-  attr_reader(*ATTRS.keys)
-
   def initialize(options = { name: 'default' })
-    resp = client.log_profile(options[:name])
-    return if resp.nil? || resp.key?('error')
-
-    @name              = resp['name']
-    @id                = resp['id']
-    @retention_policy  = resp['properties']['retentionPolicy']
-    @retention_days    = resp['properties']['retentionPolicy']['days']
-    @retention_enabled = resp['properties']['retentionPolicy']['enabled']
-
-    ATTRS.each do |name, api_name|
-      next if instance_variable_defined?("@#{name}")
-
-      instance_variable_set("@#{name}", fields[api_name])
-    end
-
-    @exists = true
-  end
-
-  def has_log_retention_enabled?
-    !!retention_enabled
-  end
-
-  def to_s
-    "#{name} Log Profile"
+    warn '[DEPRECATION] The `azure_monitor_log_profile` resource is deprecated and will be removed in version 2.0. Use the `azurerm_monitor_log_profile` resource instead.'
+    super
   end
 end
