+++ test/integration/verify/controls/azurerm_monitor_log_profile.rb
@@ -1,5 +1,5 @@
-control 'azure_monitor_activity_log_profile' do
-  describe azure_monitor_log_profile(name: 'default') do
+control 'azurerm_monitor_log_profile' do
+  describe azurerm_monitor_log_profile(name: 'default') do
     it { should exist }
     its('retention_enabled') { should be true }
     its('retention_days')    { should eq(365) }
