+++ test/integration/verify/controls/azurerm_monitor_log_profiles.rb
@@ -1,0 +1,5 @@
+control 'azurerm_monitor_log_profiles' do
+  describe azurerm_monitor_log_profiles do
+    its('names') { should include('default') }
+  end
+end
